@@ -2,42 +2,47 @@
   "name": "alis",
   "version": "1.0.0",
   "license": "Unlicense",
+  "type": "module",
+  "engines": {
+    "node": "20.x",
+    "pnpm": "8.x"
+  },
   "scripts": {
     "dev": "next dev",
     "build": "prisma generate && next build",
     "start": "next start",
     "lint": "next lint",
-    "cypress": "cypress open",
+    "cypress:open": "cypress open",
     "cypress:headless": "cypress run",
-    "e2e": "pnpm build && start-test 'pnpm start' http://0.0.0.0:3000/ 'pnpm cypress'",
-    "e2e:headless": "pnpm build && start-test 'pnpm start' http://0.0.0.0:3000/ 'pnpm cypress:headless'"
+    "e2e": "start-test 'pnpm start' http://0.0.0.0:3000/ 'pnpm cypress:open'",
+    "e2e:headless": "start-test 'pnpm start' http://0.0.0.0:3000/ 'pnpm cypress:headless'"
   },
   "prisma": {
     "schema": "src/database/schema.prisma"
   },
   "devDependencies": {
-    "@types/node": "^18.0.3",
-    "@types/nprogress": "^0.2.0",
-    "@types/react": "^18.0.15",
-    "cypress": "^10.3.0",
-    "eslint-config-silash35": "^1.0.0",
-    "prisma": "^4.0.0",
-    "sass": "^1.53.0",
-    "start-server-and-test": "^1.14.0",
-    "typescript": "^4.7.4"
+    "@types/node": "^20.11.0",
+    "@types/nprogress": "^0.2.3",
+    "@types/react": "^18.2.47",
+    "cypress": "^13.6.2",
+    "eslint-config-silash35": "^3.3.1",
+    "prisma": "^5.8.0",
+    "sass": "^1.69.7",
+    "start-server-and-test": "^2.0.3",
+    "typescript": "^5.3.3"
   },
   "dependencies": {
-    "@emotion/react": "^11.9.3",
-    "@emotion/styled": "^11.9.3",
-    "@mui/icons-material": "^5.8.4",
-    "@mui/material": "^5.8.7",
-    "@prisma/client": "^4.0.0",
-    "next": "^12.2.2",
-    "next-auth": "^4.9.0",
-    "next-pwa": "^5.5.4",
+    "@emotion/react": "^11.11.3",
+    "@emotion/styled": "^11.11.0",
+    "@mui/icons-material": "^5.15.4",
+    "@mui/material": "^5.15.4",
+    "@prisma/client": "^5.8.0",
+    "next": "^14.0.4",
+    "next-auth": "^4.24.5",
+    "next-pwa": "^5.6.0",
     "nprogress": "^0.2.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "swr": "^1.3.0"
+    "swr": "^2.2.4"
   }
 }
