@@ -3,44 +3,32 @@ import Link from "next/link";
 import styles from "./footer.module.scss";
 import ThemeSwitch from "./ThemeSwitch";
 
-const Footer = () => {
-  return (
-    <footer className={styles.footer}>
-      <div>
-        <p>
-          <Link href="/help/about">
-            <a>Sobre o ALIS</a>
-          </Link>{" "}
-          -{" "}
-          <Link href="/help/faq">
-            <a>Perguntas Frequentes</a>
-          </Link>
-        </p>
+const Footer = () => (
+  <footer className={styles.footer}>
+    <div>
+      <p>
+        <Link href="/help/about">Sobre o ALIS</Link> -{" "}
+        <Link href="/help/faq">Perguntas Frequentes</Link>
+      </p>
 
-        <p>
-          <Link href="/help/terms">
-            <a>Termos de Uso</a>
-          </Link>{" "}
-          -{" "}
-          <Link href="/help/privacy">
-            <a>Política de Privacidade</a>
-          </Link>
-        </p>
+      <p>
+        <Link href="/help/terms">Termos de Uso</Link> -{" "}
+        <Link href="/help/privacy">Política de Privacidade</Link>
+      </p>
 
-        <p>
-          Nenhum direito reservado. Mantido por{" "}
-          <a
-            href="https://silash35.github.io/"
-            title="Website of Silas Henrique"
-            rel="noopener noreferrer"
-          >
-            Silas Henrique
-          </a>
-        </p>
-      </div>
-      <ThemeSwitch />
-    </footer>
-  );
-};
+      <p>
+        Nenhum direito reservado. Mantido por{" "}
+        <a
+          href="https://silash35.github.io/"
+          rel="noopener noreferrer"
+          title="Website of Silas Henrique"
+        >
+          Silas Henrique
+        </a>
+      </p>
+    </div>
+    <ThemeSwitch />
+  </footer>
+);
 
 export default Footer;
