@@ -5,6 +5,7 @@ import Cards from "@/components/common/PlaceCards";
 import Title from "@/components/common/Title";
 import Search from "@/components/index/Search";
 
+import DeprecationWarning from "../DeprecationWarning";
 import TPlaces from "../TPlaces";
 import styles from "./main.module.scss";
 
@@ -37,9 +38,11 @@ const Main = () => {
         <p>O Agregador de Locais Inclusivos para Surdos</p>
       </Title>
 
-      <Search setSearchPlaces={setSearchPlaces} setIsSearching={setIsSearching} />
+      <Search setIsSearching={setIsSearching} setSearchPlaces={setSearchPlaces} />
 
       <Cards places={isSearching ? searchPlaces : allPlaces} />
+
+      <DeprecationWarning />
     </main>
   );
 };
