@@ -1,29 +1,31 @@
 import { GetServerSideProps, InferGetServerSidePropsType } from "next";
 import Head from "next/head";
-import { getProviders, getSession } from "next-auth/react";
+import { getServerSession } from "next-auth";
+import { getProviders } from "next-auth/react";
+import { authOptions } from "pages/api/auth/[...nextauth]";
 
 import FormLogin from "@/components/auth/FormLogin";
 
-export default function SignIn(props: InferGetServerSidePropsType<typeof getServerSideProps>) {
-  return (
-    <>
-      <Head>
-        <title>Entrar no ALIS</title>
-      </Head>
+const SignIn = (props: InferGetServerSidePropsType<typeof getServerSideProps>) => (
+  <>
+    <Head>
+      <title>Entrar no ALIS</title>
+    </Head>
 
-      <FormLogin providers={props.providers} csrfToken={props.csrfToken} />
-    </>
-  );
-}
+    <FormLogin csrfToken={props.csrfToken} providers={props.providers} />
+  </>
+);
+
+export default SignIn;
 
 export const getServerSideProps: GetServerSideProps = async (context) => {
   let csrfToken = null;
   if (process.env.ALLOW_CREDENTIALS === "true") {
-    const getCsrfToken = await import("next-auth/react").then((module) => module.getCsrfToken);
+    const { getCsrfToken } = await import("next-auth/react");
     csrfToken = await getCsrfToken(context);
   }
 
-  const session = await getSession(context);
+  const session = await getServerSession(context.req, context.res, authOptions);
   let redirect = undefined;
   if (session) {
     redirect = {
